refactor(client): split searchkeys into its own file
